docs: add `README.md` file
Create `README.md` file and write in it the name of the
project as the title and write a description for the project,
Also write `Tech Stack` section and list the used technologies
in the project.

Create `readme_file_icons` folder and add into it the used
icons (`bun.svg`, `docker.svg` and `typescript.svg`) in
`Tech Stack` section in `README.md` file.