feat: create `Validator` class
Create `Validator` class which represents a single value
(string or number) in requests inputs validation process,
This class has two properties and two methods:

- `rules` property, which is an array of validations functions
take the value as parameter and return either true (which
means the value is valid) or a string describes the reason
why the value isn't valid.

- `isOptional` property, which is a boolean value specifies
whether the value is optional or not.

- `addRule` method for adding validation rules to the value
and for using these rules later to validate the value.

- `validate` method which is used to validate the value
using the rules in `rules` array property and `isOptional`
property.