feat: create `StringValidator` class
Create `StringValidator` class for strings validation

The class uses an instance of `Validator` class to represent
the strings and validate them, it provides five string validation
rules as methods which are: `required`, `max`, `min`, `matches`
and `objectId`, these methods are used to add validation rules
to `validator` instance for using them later by `validate` method
to validate any value.

Usage:
it can be used by caling `stringValidator` function to get
an instance of `StringValidator` class and then calling
the validation methods to add validation rules (the instance
of `StringValidator` is chainable, so it is possible to chaine
multi rules in one line),  And can be validated using added
rules by calling `validate` method.