feat: create `numberValidator` class
Create `numberValidator` class for strings validation

The class uses an instance of `Validator` class to represent
the numbers and validate them, it provides three number
validation rules as methods which are: `required`, `max` and
`min`, these methods are used to add validation rules to the
`validator` instance for using them later by `validate` method
to validate any value.

Usage:
it can be used by caling `numberValidator` function to get
an instance of `NumberValidator` class and then calling
the validation methods to add validation rules (the instance
of `NumberValidator` is chainable, so it is possible to chaine
multi rules in one line),  And can be validated using the
added rules by calling `validate` method.