feat: create `validateObject` & `validationSchema` functions
Create `validateObject` function to provide a way to
validate objects using schemas, It takes the schema
definition as the 1st parameter and the object that
should be validated using the given schema as 2nd
parameter, After it validates the given object using
the given schema, The function returns the object if
it is valid, otherwise, the function will throw an error
contains the reason why the object is not valid.

Create `validationSchema` function which is just a
wrapper for `validateObject` function it is what should
be used to validate requests inputs, the function takes
the schema definition and returns an object with
`validate` method takes the object that should be
validated using that schema and calls `validateObject`
with the schema and the given object to start validation
process.