feat: datasets routes registerer
Create `src/routes/datasetsRoutes.ts` file and inside it
create `datasetsRoutesRegisterer` function which takes
the router as parameter and registers to the router the
routes or the endpoints of dealing with datasets from
application's API.

The registeration process is to call one of the HTTP
methods function of the router and pass to it the route
path as first parameter and then a list of route handlers
functions like validation middlewares and controllers to
handle the requests on this route or endpoint.
*Note* the last route handler must be a controller to
take the request to the service that should serve the
request.

For now the `datasetsRoutesRegisterer` function registers
four routes or the endpoints,

- GET /datasets => for listing all datasets.
- POST /datasets => for creating new datasets.
- DELETE /datasets => for deleting datasets.
- PATCH /datasets => for updating datasets.