feat: instructions routes registerer
Create `src/routes/instructionsRoutes.ts` file and inside
it create `instructionsRoutesRegisterer` function which
takes the router as parameter and registers to the router
the routes or the endpoints of dealing with instructions
from application's API.

The registeration process is to call one of the HTTP
methods function of the router and pass to it the route
path as first parameter and then a list of route handlers
functions like validation middlewares and controllers to
handle the requests on this route or endpoint.
*Note* the last route handler must be a controller to
take the request to the service that should serve the
request.

For now the `instructionsRoutesRegisterer` function registers
four routes or the endpoints,

- GET /instructions => for listing the instructions of a datasets.
- POST /instructions => for adding new instructions to a datasets.
- DELETE /instructions => for deleting instructions from a datasets.
- PATCH /instructions => for updating instructions in a datasets.