feat: start the HTTP server of the application
Create `src/index.ts` file and inside this file call `app`
function from `src/app.ts` to set up the application
and then call `Bun.serve` method to start the HTTP
server of the application.

And for listening to the requests to the HTTP server
of the application, pass the output of `app` function
to `fetch` option of `Bun.serve` method.

Now, the HTTP server of LLMs TDM app is ready !.