feat: create `getRecentActivities_service` service function
Create `getRecentActivities_service` for providing a way to
read the recent activities in the application.

This service function reads the recent registred activities
from the database and does some extra operations like
doing extra queries for getting the datasets details of the
registered instructions activities because it unpossible
to gather all needed information in one query.