refactor: edit the query of getting the recent activities
Edit the query of getting the recent activities in the service
function `getRecentActivities_service` and make it receives
a user id as a parameter and fetches the activities data from
user's activities document in the "recent-activities" collection
because each document in this collection represents an
activities data of a single user.

Also edit the `$lookup` stage that brings the datasets from
datasets collection to matche the new shape of datasets
collection that has been changed (by #38).