refactor: `addInstruction` service & controller functions
Make `addInstruction_service` function receives user id
as parameter from `addInstruction_controller` to be able to
use `incrementInstructionsCount` method of `datasetsService`
object to increment the instructions count.

And also make this function use `InstructionModel` as a
model directlly to add the instructions to instructions
collection instead of calling it to get the model.

Also make this function passes the user id and the data
of the created instruction to `registerInstructionActivity`
method to register the activity.