refactor: `updateInstruction` service & controller functions
Make `updateInstruction_service` function use the model
`InstructionModel` directlly to update the instructions from
instructions collection instead of calling it to get the model.

And also make this function receives user id as parameter
from `updateInstruction_controller` to pass it with the data
of the updated instructions to `registerInstructionActivity`
method that registers the activity.