refactor: `deleteInstruction` service & controller functions
Make `deleteInstruction_service` function receives user id
as parameter from `deleteInstruction_controller` to be able to
use `incrementInstructionsCount` method of `datasetsService`
object to decrement the instructions count after deleting
the instruction.

And also make this function uses `InstructionModel` as a
model directlly to delete the instructions from instructions
collection instead of calling it to get the model.

Also make this function passes the user id and the data
of the deleted instruction to `registerInstructionActivity`
method to register the activity.