refactor: `getInstructionsByIds_service` function
Make `getInstructionsByIds_service` function uses the model
`InstructionModel` directlly to get the instructions from
instructions collection instead of calling it to get the model.

Remove `datasetId` parameter because it is unnecessary to
in the query that searches by instructions ids, the dataset id
will not help to improve this but it could slow down the query
because it will add an additional check to filter the retrieved
instructions with the dataset id.