feat: create dataset instructions stream service function
Create `datasetInstructionsStream_service` function to provide
a simple and direct way to read the instructions of a specific
dataset in chunks to export the instructions of the dataset as
file (in `csv` or `jsonl` format).

The function uses `getDatasetInstructionsReader` function to
get the cursor that allows it to read the instructions from the
database group by group and it takes each group of instructions
and formats them in the requested format and calls the callback
function `onChunk` (that is passed as parapeter) with the formated
group of instructions as chunk of file.