refactor: make `initializingReplSetScript` function reusable
- Refactor `initializingReplSetScript` script function and make
it reusable and dynamic by making it imports the configuration
object that should be used to initiate a replica set by receivig
the name of the file that exports the cunfiguration object, so
the script function can be used to initiate any replica set just
by passing the name of the file that exports the cunfiguration
object of the targeted replica set to initiate as the last argument
to the process that runs `initializingReplSetScript` script function.

- Edit the way that `isItReadyMember` function checks that the
member in the replica set is ready or not.

- Also edit the log messages in `initializingReplSetScript` script
function.

- And finally move `initializingReplSetScript.js` script file into
`scripts` directory inside `mongodb` directory.