refactor: clean & rename `initializeMongodbReplSet.ts` function
Refactor `initializeMongodbReplSet.ts` script function and
make it more clear, readable by using Bun Shell ($) to
implement retrying initializing process a few times,
and make it receives an additional parameter which is
the name of the configuration file of the targeted replica
set to initialize or initiate it, the received file name will be
passed as a last argument to the process that runs
`initializingReplSetScript` script file so that `initializingReplSetScript`
script file can access this argument (which is file name) and
reads the configuration object from that file and initiates
the replica set based on that configuration object.

And finally rename `initializeMongodbReplSet.ts` script function
to `initializeReplSetOf` to reflect its reusability to initialize
any replica set of any running mongodb containers just by
passing a container name and configuration file name.