feat: add a task to `updateInstruction_service` function
Make `updateInstruction_service` function sets `isUpToDate`
property of the repository of its dataset to `false` because after
updating one of its instructions, the dataset has become not
up to date with the remote repository on huggingface hub.

Also use mongodb transaction session to update the data of
the instruction and the data of its dataset under one session
to guarantee that both operations either complete together
on fail together.