test: write one e2e test for `POST "huggingface/datasets"` route
Create `tests/e2e/huggingface/createDatasetRepository.test.ts` file
and write inside it one end to end test for `POST "huggingface/datasets"`
route.

Remove the default value of `license` property in the service function
`createDatasetRepository_service` because it is required and there is
no need to set a default value for it.

Fix a typo in `src/constants/operationsResultsMessages.ts` file.