fix: add `CORS` headers to responses for the client
Create `crosHeadersSetters` function which takes the headers
object of responses and sets the required `CORS` headers for
the client(browser) be able to receive the responses confidently
without blocking.

Use `crosHeadersSetters` function in `src/router/serveRequests.ts`
file to set CORS headers for the responses to the preflight requests
that are sent by the browsers to determine if it is safe to fetch this
resource, And for the normal responses after handling the requests.